requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.6.1
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
